+++ Package.swift
@@ -440,11 +440,6 @@ func generateMainTargets() -> [Target] {
             dependencies: [
                 "SplunkCommon",
                 "SplunkOpenTelemetry",
-                .product(
-                    name: "CrashReporter",
-                    package: "PLCrashReporter",
-                    condition: .when(platforms: [.iOS, .tvOS, .macCatalyst])
-                ),
                 resolveDependency("logger")
             ],
             path: "SplunkCustomTracking/Sources",